slight change in arg types

+++ include/display.h
@@ -5,8 +5,8 @@ namespace SCREEN
 {
     void init();
     void print(const char* message);
-    void print(int message);
-    void setCursor(int row, int col);
+    void print(uint8_t message);
+    void setCursor(uint8_t row, uint8_t col);
     void clear();
 }
